rename hintloop files to something with tiling
